Added automatic port detection for plugins. Plugins must emit 'Listening on'... followed by :port to stderr in a single line for automatic port detection to work. Repositories SELECT field names instead of * (robustness).

+++ controllers/middleware/plugins/proxy/proxy.go
@@ -14,7 +14,7 @@ import (
 
 type PluginProxyMiddleware struct {
 	Schema string
-	Port   int64
+	Port   int
 	Host   string
 }
